Added png file of xarchiver logo. Updated zip support to handle possible errors. Updated rar support with the signal XA_SUPPORT_SIGNAL_OPERATION_COMPLETE but not yet with the errors handling.Changed type of dummy_size to unsigned long long int.
(Old svn revision: 21233)

+++ libxarchiver/archive.h
@@ -77,7 +77,7 @@ struct _XAArchive
 	gchar *passwd;
 	gint nr_of_files;
 	gint nr_of_dirs;
-	unsigned long int dummy_size;
+	unsigned long long int dummy_size;
 	GSList *error_output;
 	GList *row;
 	gboolean has_passwd;
